+++ Cargo.toml
@@ -16,10 +16,17 @@ name = "nbt"
 [dependencies]
 byteorder = "1.3"
 linked-hash-map = "0.5"
-flate2 = "1.0"
+
+[dependencies.flate]
+package = "flate2"
+version = "1.0"
+optional = true
 
 [dev-dependencies]
 criterion = "0.3"
+
+[features]
+archive = ["dep:flate"]
 
 [[bench]]
 name = "decode"
